pyproject.toml
[tool.poetry]
name = "python-urllib"
version = "0.1.0"
description = ""
authors = ["José Guilherme Fernandes <josegfernandes03@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pywhatkit = "^5.4"
python-dotenv = "^0.21.0"


[tool.poetry.group.dev.dependencies]
black = {version = "^22.8.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
